Rename parameter estimator file
