+++ gammapy/estimators/parameter.py
